third_party/blink/web_tests/external/wpt/trusted-types/require-trusted-types-for-expected.txt

This is a testharness.js-based test.
[FAIL] Set require trusted types for 'script' without CSP for trusted types don't block policy creation and using.
  assert_equals: expected "HTMLScriptElement innerText" but got "HTMLElement innerText"
Harness: the test ran to completion.
